@@ -1,5 +1,4 @@
 ﻿@using Microsoft.AspNetCore.Identity
 @using MVCStore.Areas.Identity
-@using Microsoft.AspNetCore.Identity
-@namespace MVCStore.Areas.Identity.Pages
+@using MVCStore.Areas.Identity.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
