@@ -2,14 +2,17 @@
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 using MVCStore.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 
 namespace MVCStore.Data
 {
-    public class ApplicationDbContext : IdentityDbContext<IdentityUser>
-    {
-        public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options)
-            : base(options) { }
+	public class ApplicationDbContext : IdentityDbContext<IdentityUser>
+	{
+		public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options) : base(options) { }
 
-        public DbSet<Product> Products { get; set; }
-    }
+		public DbSet<Product> Products { get; set; }
+	}
 }
