@@ -1,8 +1,14 @@
 ﻿@model IEnumerable<Product>
+
 @{
     ViewBag.Title = "All Products";
     Layout = "_AdminLayout";
 }
+
+<h1>Products</h1>
+
+<a asp-action="Create" class="btn btn-primary mb-3">Add Product</a>
+
 <table class="table table-striped table-bordered table-sm">
     <tr>
         <th class="text-right">ID</th>
@@ -15,14 +21,14 @@
         <tr>
             <td class="text-right">@item.ProductID</td>
             <td>@item.Name</td>
-            <td class="text-right">@item.Price.ToString("c")</td>
+            <td class="text-right">@item.Price</td>
             <td class="text-center">
-                <form asp-action="Delete" method="post">
-                    <a asp-action="Edit" class="btn btn-sm btn-warning"
-                       asp-route-productId="@item.ProductID">
-                        Edit
-                    </a>
-                    <input type="hidden" name="ProductID" value="@item.ProductID" />
+                <a asp-action="Edit" class="btn btn-sm btn-warning"
+                   asp-route-productId="@item.ProductID">
+                    Edit
+                </a>
+                <form asp-action="Delete" method="post" style="display: inline">
+                    <input type="hidden" name="ProductId" value="@item.ProductID" />
                     <button type="submit" class="btn btn-danger btn-sm">
                         Delete
                     </button>
@@ -31,6 +37,3 @@
         </tr>
     }
 </table>
-<div class="text-center">
-    <a asp-action="Create" class="btn btn-primary">Add Product</a>
-</div>
