@@ -1,2 +1,4 @@
-﻿@using MVCStore.Models
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+﻿@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+@addTagHelper *, MVCStore
+@using MVCStore.Models
+@using MVCStore.ViewModels
